refactor: apply /simplify findings on vm logs
- Fold LogFilePath/LogPath into hypervisor/helpers.go alongside
  PIDFilePath / SocketPath (matches sibling-style: same shape, same
  Backend method, all per-VM path builders live in one file).
- Replace tailFollow polling with utils.WatchFile (fsnotify, already
  used by vm status). 100ms debounce, exits on ctx.Done() or watcher
  channel close. Eliminates ~120 idle read syscalls/min.
- Reword the os.IsNotExist error to lead with the action verb
  ("open log %s: VM may not have been started yet"), matching the
  rest of cocoon's error format.
